llm_backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
deepface==0.0.93
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0
pymongo==4.8.0
python-dotenv==1.0.1
openai==1.50.2
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
tf-keras==2.15.0
httpx==0.27.2
httpcore==1.0.5
anyio==4.4.0